+++ CHANGELOG.md
@@ -1,11 +1,27 @@
+<!--
+ * @Author: haobin.wang
+ * @Date: 2024-12-17 11:43:12
+ * @LastEditors: haobin.wang
+ * @LastEditTime: 2025-01-03 09:44:46
+ * @Description: Do not edit
+-->
 # Change Log
 
 ### Added
 - Pag File Preview: A Visual Studio Code extension for previewing PAG files with ease and efficiency.
 - Documentation updates.
 
+## [1.0.4] - 2025-01-03
 ### Fixed
--
+- Fixed the issue that webview keeps the session window from playing
+
+## [1.0.3] - 2025-01-02
+### Changed
+- Downgrade the supported version to 1.87.0
+
+## [1.0.2] - 2024-12-25
+### Added
+- Increase file size display
 
 ## [1.0.0] - 2024-12-18
 ### Added